refactor safeParse
